Crate azalea_entity

Re-exports

pub use attributes::Attributes;

Modules

attributes

See https://minecraft.fandom.com/wiki/Attribute.

indexing

Stuff related to entity indexes and keeping track of entities in the world.

metadata

mining

particle

Structs

Dead

Marker component for entities that are dead.

EntityBundle

A bundle of components that every entity has. This doesn’t contain metadata, that has to be added separately.

EntityDataItem

EntityDimensions

EntityKind

A component NewType for azalea_registry::EntityKind.

EntityMetadataItems

EntityPlugin

Plugin handling some basic entity functionality.

EntityUuid

The Minecraft UUID of the entity. For players, this is their actual player UUID, and for other entities it’s just random.

EyeHeight

A component that contains the offset of the entity’s eyes from the entity coordinates.

FluidOnEyes

InLoadedChunk

Marks an entity that’s in a loaded chunk. This is updated at the beginning of every tick.

Jumping

A component for entities that can jump.

LastSentPosition

The last position of the entity that was sent over the network.

LoadedBy

A component that lists all the local player entities that have this entity loaded. If this is empty, the entity will be removed from the ECS.

LocalEntity

A marker component that signifies that this entity is “local” and shouldn’t be updated by other clients.

LookDirection

A component that contains the direction an entity is looking.

OnClimbable

OptionalUnsignedInt

Physics

The physics data relating to the entity, such as position, velocity, and bounding box.

PlayerBundle

A bundle of the components that are always present for a player.

Position

The position of the entity right now.

Quaternion

RelativeEntityUpdate

An [EntityCommand] that applies a “relative update” to an entity, which means this update won’t be run multiple times by different clients in the same world.

Rotations

A set of x, y, and z rotations. This is used for armor stands.

VillagerData

Enums

ArmadilloStateKind

EntityDataValue

EntityUpdateSet

A Bevy [SystemSet] for various types of entity updates.

Pose

SnifferState

Functions

add_dead

System that adds the Dead marker component if an entity’s health is set to 0 (or less than 0). This will be present if an entity is doing the death animation.

clamp_look_direction

input_vector

move_relative

on_pos

on_pos_legacy

Get the position of the block below the entity, but a little lower.

update_bounding_box

Sets the position of the entity. This doesn’t update the cache in azalea-world, and should only be used within azalea-world!

update_fluid_on_eyes

update_in_loaded_chunk

Update the InLoadedChunk component for all entities in the world.

update_on_climbable

view_vector