azalea_entity
Module indexing

Stuff related to entity indexes and keeping track of entities in the world.

Structs

EntityChunkPos

The chunk position that an entity is currently in.

EntityIdIndex

An index of Minecraft entity IDs to Azalea ECS entities. This is a Component so local players can keep track of entity IDs independently from the instance.

EntityUuidIndex

Functions

add_entity_to_indexes

insert_entity_chunk_position

Insert new entities into Instance::entities_by_chunk.

remove_despawned_entities_from_indexes

Despawn entities that aren’t being loaded by anything.

update_entity_chunk_positions

Update the chunk position indexes in Instance::entities_by_chunk.