azalea_entity
Struct LoadedBy

pub struct LoadedBy(pub HashSet<Entity>);

A component that lists all the local player entities that have this entity loaded. If this is empty, the entity will be removed from the ECS.

Tuple Fields
0: HashSet<Entity>